Refactored analytics code into multiple files.
